+++ typings-checker/index.ts
@@ -146,16 +146,6 @@ x27[0] = 2
 x27.push(2)
 
 //
-// fold on unions
-//
-
-const f1 = t.union([t.string, t.number]).fold(s => s.length, n => n)
-f1('a')
-f1(1)
-// $ExpectError Argument of type 'true' is not assignable to parameter of type 'string | number'
-f1(true)
-
-//
 // map
 //